[Feature] Charting Modularity (#6477)
* add accessors to the charting extension

* equity views

* get charting functions from the extensions/accessors

* use the private variable instead

* etf views

* index views

* crypto and currency views

* technical views

* economy views

* fixedincome views

* right description for price historical

* unused

* adjust test and deprecate the charting router

* import views only with charting installed

* adding views throught dependencies instead

* removing unnecessary infra and adding entry point to charting extension

* fixed income toml

* removing accessor

* typo

* typo

* adding extra params to obbject

* remove things from init

* rename utils to charts instead

* updated readme

---------
