+++ openbb_platform/extensions/currency/openbb_currency/currency_views.py
@@ -1,0 +1,17 @@
+"""Views for the Currency Extension."""
+
+from typing import Any, Dict, Tuple
+
+from openbb_charting.charts.price_historical import price_historical
+from openbb_charting.core.openbb_figure import OpenBBFigure
+
+
+class CurrencyViews:
+    """Currency Views."""
+
+    @staticmethod
+    def currency_price_historical(  # noqa: PLR0912
+        **kwargs,
+    ) -> Tuple[OpenBBFigure, Dict[str, Any]]:
+        """Currency Price Historical Chart."""
+        return price_historical(**kwargs)
