+++ openbb_platform/extensions/currency/pyproject.toml
@@ -17,3 +17,6 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.plugins."openbb_core_extension"]
 currency = "openbb_currency.currency_router:router"
+
+[tool.poetry.plugins."openbb_charting_extension"]
+currency = "openbb_currency.currency_views:CurrencyViews"
