+++ openbb_platform/extensions/fixedincome/pyproject.toml
@@ -17,3 +17,6 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.plugins."openbb_core_extension"]
 fixedincome = "openbb_fixedincome.fixedincome_router:router"
+
+[tool.poetry.plugins."openbb_charting_extension"]
+fixedincome = "openbb_fixedincome.fixedincome_views:FixedIncomeViews"
